requirements.txt
torch>=2.5.0
torchvision
torchaudio
transformers>=4.40.0
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pyyaml>=6.0

# Optional — only needed for source: s3 in models.yaml
# boto3>=1.34.0
